requirements.txt
pdfplumber
langchain-text-splitters
sentence-transformers
faiss-cpu
transformers
torch
gradio
matplotlib
numpy